# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Weave Control Plane API
  version: 0.1.0
  description: |
    The Weave control-plane HTTP API.

    This document is the source of truth for the API surface and is updated
    before or alongside the implementation it describes. The TypeScript client
    is generated from it; response types are never hand-written in the web
    application.

    Only the authenticated user's own profile is defined so far. Workspaces,
    tasks and sessions arrive in later units.

    Operational probes (`/health/live`, `/health/ready`) are deliberately not
    described here: they are an operator interface, not part of the versioned
    product API, and their shape is free to change with the deployment.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /v1/workspaces/{workspaceId}/invitations:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      operationId: listWorkspaceInvitations
      tags: [workspaces]
      summary: List a workspace's invitations
      description: Requires `member:invite`. Never returns tokens or token hashes.
      responses:
        "200":
          description: Every invitation for the workspace, newest first.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Invitation" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PermissionDenied" }
        "404": { $ref: "#/components/responses/WorkspaceNotFound" }
    post:
      operationId: issueWorkspaceInvitation
      tags: [workspaces]
      summary: Invite someone to the workspace
      description: |
        Requires `member:invite`, and the inviter must be able to grant the
        role — an admin cannot invite an owner, for the same reason an admin
        cannot promote one.

        **The response carries the token, and it is the only time it ever
        exists outside the recipient's link.** Only its hash is stored, so it
        cannot be retrieved afterwards. Surface it immediately or it is lost.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, role]
              properties:
                email: { type: string, format: email }
                role: { $ref: "#/components/schemas/Role" }
      responses:
        "201":
          description: The invitation, together with its token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IssuedInvitation" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PermissionDenied" }
        "404": { $ref: "#/components/responses/WorkspaceNotFound" }
        "409":
          description: The address already belongs to the workspace, or already has an outstanding invitation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/workspaces/{workspaceId}/invitations/{invitationId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - name: invitationId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: revokeWorkspaceInvitation
      tags: [workspaces]
      summary: Withdraw an invitation
      description: Requires `member:invite`. Only an outstanding invitation can be revoked.
      responses:
        "204": { description: The invitation was withdrawn. }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PermissionDenied" }
        "404": { $ref: "#/components/responses/WorkspaceNotFound" }
  /v1/invitations/preview:
    post:
      operationId: previewInvitation
      tags: [invitations]
      summary: Describe an invitation to its holder
      description: |
        Lets someone see what they are being asked to join before accepting.

        `POST` rather than `GET` because the token travels in the body: a
        token in a query string ends up in server logs, proxy logs and
        `Referer` headers.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationTokenRequest" }
      responses:
        "200":
          description: What the invitation offers, and nothing else about the workspace.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationPreview" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/InvitationNotUsable" }
  /v1/invitations/accept:
    post:
      operationId: acceptInvitation
      tags: [invitations]
      summary: Accept an invitation
      description: |
        Requires authentication, but not membership — the caller is not a
        member yet, and the token is what authorizes them.

        The signed-in account's email must match the invited address. A link
        that admits whoever opens it would turn a forwarded message into a
        workspace breach.

        The token is single use: concurrent accepts of one token produce
        exactly one membership.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvitationTokenRequest" }
      responses:
        "200":
          description: The membership created by accepting.
          content:
            application/json:
              schema:
                type: object
                required: [workspace_id, role]
                properties:
                  workspace_id: { type: string, format: uuid }
                  role: { $ref: "#/components/schemas/Role" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403":
          description: The invitation was issued to a different email address.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/InvitationNotUsable" }
        "409":
          description: The caller already belongs to this workspace.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A session token issued by the configured identity provider, presented
        as `Authorization: Bearer <token>`.
  schemas:
    Role:
      type: string
      description: |
        A member's position in a workspace. Permissions are derived from the
        role by the server; clients read `permissions` rather than mapping
        roles themselves.
      enum: [owner, admin, developer, viewer]
...
    Invitation:
      type: object
      additionalProperties: false
      required: [id, email, role, status, expires_at, created_at]
      description: An offer to join a workspace. Never carries the token or its hash.
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        status:
          type: string
          description: |
            Derived from the invitation's timestamps rather than stored, so it
            cannot disagree with them. A terminal state wins over expiry: an
            invitation accepted while still valid reads as `accepted`.
          enum: [pending, accepted, revoked, expired]
        invited_by_email: { type: string, format: email }
        invited_by_display_name: { type: string }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    IssuedInvitation:
      allOf:
        - $ref: "#/components/schemas/Invitation"
        - type: object
          required: [token]
          properties:
            token:
              type: string
              description: |
                **Shown once.** Only a SHA-256 hash is stored, so this value
                cannot be recovered afterwards — not by this API, and not by
                anyone with database access. Put it in the invitation link
                immediately.
    InvitationTokenRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token: { type: string }
    InvitationPreview:
      type: object
      additionalProperties: false
      required: [workspace_name, email, role, invited_by_email, expires_at]
      description: |
        Deliberately narrow. It tells the holder what they are joining and who
        asked, and reveals nothing else about the workspace or its members.
      properties:
        workspace_name: { type: string }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        invited_by_email: { type: string, format: email }
        invited_by_display_name: { type: string }
        expires_at: { type: string, format: date-time }
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      description: |
        The error shape returned by every failing endpoint.

        It never contains a stack trace, database error, provider response or
        internal hostname. Quote `request_id` when reporting a problem: it
        correlates to the server-side logs that hold the diagnosis.
      properties:
        code:
          type: string
          description: Stable, machine-readable identifier. Clients branch on this rather than on `message`.
          enum:
            - unauthenticated
            - permission_denied
            - not_found
            - conflict
            - invalid_request
            - internal_error
        message:
          type: string
          description: Human-readable summary, safe to surface to an end user.
        request_id:
          type: string
          description: Correlates this response with server-side logs. Also returned in the `X-Request-Id` header.
  responses:
    Unauthenticated:
      description: |
        No valid credential was presented.

        Returned identically whether the token was absent, malformed, expired,
        signed by the wrong key, or issued for a different audience.
        Distinguishing them would tell a caller how close a forgery attempt
        came, and no legitimate client needs the difference.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: unauthenticated
            message: Authentication required. Present a valid bearer token.
            request_id: 0199a3b4-1c2d-7e3f-8a9b-0c1d2e3f4a5b
    WorkspaceNotFound:
      description: |
        No such workspace, **or** the caller is not a member of it. The two are
        deliberately indistinguishable: answering `403` for a workspace that
        exists but is not the caller's would let anyone enumerate tenants by
        guessing identifiers.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PermissionDenied:
      description: |
        The caller is a member of this workspace but their role does not hold
        the required permission. Distinct from `404`, which hides the workspace
        entirely.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    InvitationNotUsable:
      description: |
        The invitation cannot be used.

        Returned identically whether the token is unknown, expired, revoked or
        already accepted. Distinguishing them would tell someone guessing
        tokens which guesses were once real, turning a brute-force attempt
        into a search with feedback.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InvalidRequest:
      description: The body was malformed, carried unknown fields, or failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...